Modify autosave logic so that backend and frontend perform the same checks
This modifies the checks that are carried out to determine whether a post
is autosaveable. Previously, the frontend carried out a comparison
between the current autosave and the proposed autosave. If there were
differences, the post would be saved. The backend, however, carried out
checks between the latest post content and the proposed autosave.

Because the frontend and backend checks were different failures could
occur on the backend in certain conditions (when the post is different
to the latest autosave, but the same as the current post).

This commit ensures the frontend makes the same checks as the backend.

The proposed autosave is checked first against post content. If different,
the proposed autosave is next checked against the most recent autosave.